+++ CHANGELOG.md
@@ -13,6 +13,35 @@ Change log entries should be formulated in the imperative present tense.
 
 ## Unreleased
 
+
+## [1.0.0] - 2018-12-14
+
+### Changed
+
+* Change the `TestProbe` parameter in `AkkaUnorderedExpectations.CreateExpectedMarbleForUnorderedGroup` to be of type `Action<Action<T>>`.
+* Change the type of the `assertionFactory` parameter from `Func<string, Action<T>>` to `Action<string, T>` in `MarbleTestExtensionsForAkka.ExpectMsgs` overloads.
+* The following types are now `internal`:
+    * `ExpectedMarble`
+    * `ExpectedMarbles`
+    * `InputMarble`
+    * `Moment`
+* The accessibility of the lists `Expectations` and `Inputs` on `MarbleTest` is now `private`.
+* The accessibility of the methods `CreateInputMarbles` and `ParseSequence` is now `private`.
+* Change different constructors of `Moment` to static factory methods.
+* Change the `bool` `IsOrderedGroup` property to a `MomentType` `enum` property `Type`.
+
+### Added
+
+* Add methods `Assert` and `ExpectAtLeast` to `MarbleTest`
+* Add `TimeShiftedClone` method to `Moment`.
+
+### Deleted
+
+* Remove `[Obsolete(...)]` `ExpectMsgs` overloads that use predicates.
+* Remove `MarbleParser`, the `MultiCharMarbleParser` is the only parser that can be used.
+* Remove `MarbleTest` constructors that allow specifying the marble parser to use.
+
+
 ## [0.7.0] - 2018-11-22
 
 ### Added
@@ -67,7 +96,7 @@ Change log entries should be formulated in the imperative present tense.
 
 ### Added
 
-* Support for multi character marbles. 
+* Support for multi character marbles.
 
 ## [0.1.0] - 2016-11-28